+++ src/js/components/recipient/RecipientPage.jsx
@@ -3,20 +3,23 @@
  * Created by Lizzie Salita 8/23/17
  */
 
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { ShareIcon, FiscalYearPicker } from 'data-transparency-ui';
-
+import { find, throttle } from 'lodash';
+import { useHistory } from "react-router-dom";
+import { useQueryParams } from 'helpers/queryParams';
+import { scrollToY } from 'helpers/scrollToHelper';
 import { currentFiscalYear, earliestFiscalYear, getFiscalYearsWithLatestAndAll } from 'helpers/fiscalYearHelper';
 import { recipientPageMetaTags } from 'helpers/metaTagHelper';
-
 import { LoadingWrapper } from "components/sharedComponents/Loading";
 import { getBaseUrl, handleShareOptionClick } from 'helpers/socialShare';
-
+import { stickyHeaderHeight } from 'dataMapping/stickyHeader/stickyHeader';
 import ChildRecipientModalContainer from 'containers/recipient/modal/ChildRecipientModalContainer';
 import { AlternateNamesRecipientModalContainer } from 'containers/recipient/modal/AlternateNamesRecipientModalContainer';
 import PageWrapper from 'components/sharedComponents/PageWrapper';
 import Error from 'components/sharedComponents/Error';
+import { getStickyBreakPointForSidebar } from 'helpers/stickyHeaderHelper';
 
 import RecipientContent from './RecipientContent';
 
@@ -35,9 +38,11 @@ export const RecipientPage = ({
     error,
     pickedFy
 }) => {
+    const history = useHistory();
+    const query = useQueryParams();
     const [isChildModalVisible, showChildModal] = useState(false);
     const [isAlternateModalVisible, showAlternateRecipientModal] = useState(false);
-
+    const [activeSection, setActiveSection] = useState(query.section || 'overview');
     const showAlternateModal = () => showAlternateRecipientModal(true);
     const hideAlternateModal = () => showAlternateRecipientModal(false);
     const showChildRecipientModal = () => showChildModal(true);
@@ -52,6 +57,57 @@ export const RecipientPage = ({
     const handleShare = (name) => {
         handleShareOptionClick(name, slug, emailArgs);
     };
+
+    const recipientSections = [
+        {
+            section: 'overview',
+            label: 'Overview'
+        },
+        {
+            section: 'transactions-over-time',
+            label: 'Transactions Over Time'
+        },
+        {
+            section: 'top-five',
+            label: 'Top 5'
+        }
+    ];
+
+    const jumpToSection = (section = '') => {
+        // we've been provided a section to jump to
+        // check if it's a valid section
+        const sectionObj = find(recipientSections, ['label', section]);
+        if (!sectionObj) return;
+
+        // find the section in dom
+        const sectionDom = document.querySelector(`#recipient-${sectionObj.section}`);
+        if (!sectionDom) return;
+
+        // add section to url
+        history.replace(`?section=${sectionObj.section}`);
+
+        // add offsets
+        const conditionalOffset = window.scrollY < getStickyBreakPointForSidebar() ? stickyHeaderHeight : 10;
+        const sectionTop = (sectionDom.offsetTop - stickyHeaderHeight - conditionalOffset);
+        scrollToY(sectionTop - 25, 700);
+    };
+
+    useEffect(throttle(() => {
+        // this allows the page to jump to a section on page load, when
+        // using a link to open the page
+        // prevents a console error about react unmounted component leak
+        let isMounted = true;
+        if (isMounted) {
+            const urlSection = query.section;
+            if (urlSection) {
+                setActiveSection(urlSection);
+                jumpToSection(urlSection);
+            }
+        }
+        return () => {
+            isMounted = false;
+        };
+    }, 100), [history, query.section]);
 
     let content = (
         <RecipientContent
@@ -72,6 +128,7 @@ export const RecipientPage = ({
         backgroundColor: "#1a4480"
     };
 
+
     return (
         <PageWrapper
             pageName="Recipient Profile"
@@ -88,7 +145,10 @@ export const RecipientPage = ({
                 <ShareIcon
                     onShareOptionClick={handleShare}
                     url={getBaseUrl(slug)} />
-            ]}>
+            ]}
+            sections={recipientSections}
+            activeSection={activeSection}
+            jumpToSection={jumpToSection}>
             <>
                 <main id="main-content" className="main-content">
                     <LoadingWrapper isLoading={loading}>
